refactor: flatten blocks/ subfolders and move Footer into its own directory
- Flatten blocks/Animations/ and blocks/TextAnimations/ (remove
  redundant single-file subfolders like BlurFade/BlurFade.tsx)
- Move Footer.tsx into components/Footer/ for consistency
- Update all import paths accordingly
