+++ src/blocks/Animations/BlurFade.tsx
