+++ src/blocks/Animations/FlickeringGrid.tsx
