+++ src/blocks/Animations/HoverBorderGradient.tsx
@@ -1,6 +1,6 @@
 import { motion } from 'framer-motion';
 import { type ReactNode, useCallback, useEffect, useState } from 'react';
-import { cn } from '../../../lib/utils';
+import { cn } from '../../lib/utils';
 
 type Direction = 'TOP' | 'LEFT' | 'BOTTOM' | 'RIGHT';
