+++ src/blocks/TextAnimations/CountUp.tsx
