+++ src/blocks/TextAnimations/SplitText.tsx
